Updates Made in this commit: .) MealManager.java only updates Time value on update
Next To Do:
.) Empty Pie chart has weird values

Previous Updates:
.)

Errors
.)

+++ src/main/java/App_Code/Objects/Database_Objects/MealManagerRegistry.java
@@ -128,7 +128,7 @@ public class MealManagerRegistry
             String macroName = mapEntry.getKey();
             Integer macroPos = mapEntry.getValue().getValue0();
             
-            BigDecimal macroValue = totalMealTable.get_ValueOnTable(0, macroPos);
+            BigDecimal macroValue = (BigDecimal) totalMealTable.get_ValueOnTable(0, macroPos);
             
             /**
              *  <Key: MacroName | Value: HashMap <Key: MealManager, Value:  Quantity>>
